extract state to hook
